chore: stage and push all recent changes
