+++ tmp/unused/src/core/functions/get/pyside6_qt_get_shot_name.py
