+++ tmp/unused/src/core/functions/open/pyside6_qt_open_in_finder.py
