+++ tmp/unused/src/core/functions/resolve/pyside6_qt_resolve_shot_name.py
