+++ tmp/unused/src/core/ui/orchestrators/pyside6_qt_flame_functions.py
@@ -2,7 +2,7 @@
 
 # -------------------------------------------------------------------------- #
 
-# File Name:        pyside6_qt_flame_classes.py
+# File Name:        pyside6_qt_flame_functions.py
 # Version:          1.1.0
 # Created:          2024-01-19
 # Modified:         2025-11-15
@@ -21,26 +21,19 @@ if src_parent not in sys.path:
     sys.path.insert(0, src_parent)
 
 # -------------------------------------------------------------------------- #
-# Imports - Widget Classes
+# Imports - Widget Functions
 # -------------------------------------------------------------------------- #
 
-from tmp.unused.src.ui.widgets.pyside6_qt_button import pyside6_qt_button
-from tmp.unused.src.ui.widgets.pyside6_qt_clickable_line_edit import pyside6_qt_clickable_line_edit
-from tmp.unused.src.ui.widgets.pyside6_qt_label import pyside6_qt_label
-from tmp.unused.src.ui.widgets.pyside6_qt_line_edit import pyside6_qt_line_edit
-from tmp.unused.src.ui.widgets.pyside6_qt_list_widget import pyside6_qt_list_widget
-from tmp.unused.src.ui.dialog.pyside6_qt_message_window import pyside6_qt_message_window
-from tmp.unused.src.ui.dialog.pyside6_qt_password_window import pyside6_qt_password_window
-from tmp.unused.src.ui.dialog.pyside6_qt_preset_window import pyside6_qt_preset_window
-from tmp.unused.src.ui.dialog.pyside6_qt_progress_window import pyside6_qt_progress_window
-from tmp.unused.src.ui.widgets.pyside6_qt_push_button import pyside6_qt_push_button
-from tmp.unused.src.ui.widgets.pyside6_qt_push_button_menu import pyside6_qt_push_button_menu
-from tmp.unused.src.ui.dialog.pyside6_qt_qdialog import pyside6_qt_qdialog
-from tmp.unused.src.ui.widgets.pyside6_qt_slider import pyside6_qt_slider
-from tmp.unused.src.ui.widgets.pyside6_qt_text_edit import pyside6_qt_text_edit
-from tmp.unused.src.ui.widgets.pyside6_qt_token_push_button import pyside6_qt_token_push_button
-from tmp.unused.src.ui.widgets.pyside6_qt_tree_widget import pyside6_qt_tree_widget
-from tmp.unused.src.ui.widgets.pyside6_qt_window import pyside6_qt_window
+from tmp.unused.src.core.functions.browse.pyside6_qt_file_browser import pyside6_qt_file_browser
+from tmp.unused.src.core.functions.get.pyside6_qt_get_flame_version import pyside6_qt_get_flame_version
+from tmp.unused.src.core.functions.get.pyside6_qt_get_shot_name import pyside6_qt_get_shot_name
+from tmp.unused.src.core.functions.io.pyside6_qt_load_config import pyside6_qt_load_config
+from tmp.unused.src.core.functions.open.pyside6_qt_open_in_finder import pyside6_qt_open_in_finder
+from tmp.unused.src.core.functions.print.pyside6_qt_print import pyside6_qt_print
+from tmp.unused.src.core.functions.refresh.pyside6_qt_refresh_hooks import pyside6_qt_refresh_hooks
+from tmp.unused.src.core.functions.resolve.pyside6_qt_resolve_path_tokens import pyside6_qt_resolve_path_tokens
+from tmp.unused.src.core.functions.resolve.pyside6_qt_resolve_shot_name import pyside6_qt_resolve_shot_name
+from tmp.unused.src.core.functions.io.pyside6_qt_save_config import pyside6_qt_save_config
 
 
 # -------------------------------------------------------------------------- #
