+++ tmp/unused/src/ui/pyside6_qt_flame_modules.py
@@ -34,44 +34,44 @@ except ImportError:
 # Imports - LOGIK-PROJEKT UI Classes (Direct from widgets)
 # ========================================================================== #
 
-from tmp.unused.src.ui.widgets.pyside6_qt_button import pyside6_qt_button
-from tmp.unused.src.ui.widgets.pyside6_qt_clickable_line_edit import pyside6_qt_clickable_line_edit
-from tmp.unused.src.ui.widgets.pyside6_qt_label import pyside6_qt_label
-from tmp.unused.src.ui.widgets.pyside6_qt_line_edit import pyside6_qt_line_edit
-from tmp.unused.src.ui.widgets.pyside6_qt_list_widget import pyside6_qt_list_widget
-from tmp.unused.src.ui.dialog.pyside6_qt_message_window import pyside6_qt_message_window
-from tmp.unused.src.ui.dialog.pyside6_qt_password_window import pyside6_qt_password_window
-from tmp.unused.src.ui.dialog.pyside6_qt_preset_window import pyside6_qt_preset_window
-from tmp.unused.src.ui.dialog.pyside6_qt_progress_window import pyside6_qt_progress_window
-from tmp.unused.src.ui.widgets.pyside6_qt_push_button import pyside6_qt_push_button
-from tmp.unused.src.ui.widgets.pyside6_qt_push_button_menu import pyside6_qt_push_button_menu
-from tmp.unused.src.ui.dialog.pyside6_qt_qdialog import pyside6_qt_qdialog
-from tmp.unused.src.ui.widgets.pyside6_qt_slider import pyside6_qt_slider
-from tmp.unused.src.ui.widgets.pyside6_qt_text_edit import pyside6_qt_text_edit
-from tmp.unused.src.ui.widgets.pyside6_qt_token_push_button import pyside6_qt_token_push_button
-from tmp.unused.src.ui.widgets.pyside6_qt_tree_widget import pyside6_qt_tree_widget
-from tmp.unused.src.ui.widgets.pyside6_qt_window import pyside6_qt_window
+from tmp.unused.src.core.ui.widgets.button.pyside6_qt_button import pyside6_qt_button
+from tmp.unused.src.core.ui.widgets.line_edit.pyside6_qt_clickable_line_edit import pyside6_qt_clickable_line_edit
+from tmp.unused.src.ui.widgets.label.pyside6_qt_label import pyside6_qt_label
+from tmp.unused.src.core.ui.widgets.line_edit.pyside6_qt_line_edit import pyside6_qt_line_edit
+from tmp.unused.src.core.ui.widgets.list.pyside6_qt_list_widget import pyside6_qt_list_widget
+from tmp.unused.src.core.ui.dialog.pyside6_qt_message_window import pyside6_qt_message_window
+from tmp.unused.src.core.ui.dialog.pyside6_qt_password_window import pyside6_qt_password_window
+from tmp.unused.src.core.ui.dialog.pyside6_qt_preset_window import pyside6_qt_preset_window
+from tmp.unused.src.core.ui.dialog.pyside6_qt_progress_window import pyside6_qt_progress_window
+from tmp.unused.src.core.ui.widgets.button.pyside6_qt_push_button import pyside6_qt_push_button
+from tmp.unused.src.core.ui.widgets.button.pyside6_qt_push_button_menu import pyside6_qt_push_button_menu
+from tmp.unused.src.core.ui.dialog.pyside6_qt_qdialog import pyside6_qt_qdialog
+from tmp.unused.src.core.ui.widgets.slider.pyside6_qt_slider import pyside6_qt_slider
+from tmp.unused.src.core.ui.widgets.text_edit.pyside6_qt_text_edit import pyside6_qt_text_edit
+from tmp.unused.src.core.ui.widgets.button.pyside6_qt_token_push_button import pyside6_qt_token_push_button
+from tmp.unused.src.core.ui.widgets.tree.pyside6_qt_tree_widget import pyside6_qt_tree_widget
+from tmp.unused.src.core.ui.widgets.window.pyside6_qt_window import pyside6_qt_window
 
 # ========================================================================== #
 # Imports - LOGIK-PROJEKT UI Functions (Direct from widgets)
 # ========================================================================== #
 
-from tmp.unused.src.utils.pyside6_qt_file_browser import pyside6_qt_file_browser
-from tmp.unused.src.utils.pyside6_qt_get_flame_version import pyside6_qt_get_flame_version
-from tmp.unused.src.utils.pyside6_qt_get_shot_name import pyside6_qt_get_shot_name
-from tmp.unused.src.utils.io.pyside6_qt_load_config import pyside6_qt_load_config
-from tmp.unused.src.utils.pyside6_qt_open_in_finder import pyside6_qt_open_in_finder
+from tmp.unused.src.core.functions.browse.pyside6_qt_file_browser import pyside6_qt_file_browser
+from tmp.unused.src.core.functions.get.pyside6_qt_get_flame_version import pyside6_qt_get_flame_version
+from tmp.unused.src.core.functions.get.pyside6_qt_get_shot_name import pyside6_qt_get_shot_name
+from tmp.unused.src.core.functions.io.pyside6_qt_load_config import pyside6_qt_load_config
+from tmp.unused.src.core.functions.open.pyside6_qt_open_in_finder import pyside6_qt_open_in_finder
 from tmp.unused.src.utils.pyside6_qt_print import pyside6_qt_print
-from tmp.unused.src.utils.pyside6_qt_refresh_hooks import pyside6_qt_refresh_hooks
-from tmp.unused.src.utils.pyside6_qt_resolve_path_tokens import pyside6_qt_resolve_path_tokens
-from tmp.unused.src.utils.pyside6_qt_resolve_shot_name import pyside6_qt_resolve_shot_name
+from tmp.unused.src.core.functions.refresh.pyside6_qt_refresh_hooks import pyside6_qt_refresh_hooks
+from tmp.unused.src.core.functions.resolve.pyside6_qt_resolve_path_tokens import pyside6_qt_resolve_path_tokens
+from tmp.unused.src.core.functions.resolve.pyside6_qt_resolve_shot_name import pyside6_qt_resolve_shot_name
 from tmp.unused.src.utils.io.pyside6_qt_save_config import pyside6_qt_save_config
 
 # ========================================================================== #
 # Imports - LOGIK-PROJEKT UI Components
 # ========================================================================== #
 
-from src.ui.methods.pyside6_qt_output_config_ui import (
+from tmp.unused.src.core.ui.layouts.pyside6_qt_output_config_ui import (
     pyside6_qt_output_config_ui
 )
